Some fixes

+++ WowPacketParserModule.V7_2_0_23758/Hotfix/CinematicSequencesEntry.cs
@@ -6,7 +6,7 @@ namespace WowPacketParserModule.V7_2_0_23758.Hotfix
     [HotfixStructure(DB2Hash.CinematicSequences, HasIndexInData = false)]
     public class CinematicSequencesEntry
     {
-        public ushort SoundID { get; set; }
+        public uint SoundID { get; set; }
         [HotfixArray(8)]
         public ushort[] Camera { get; set; }
     }
